requirements.txt
google-generativeai==0.3.1
openai==1.14.0
python-dotenv==1.0.0
numpy==1.24.4
pandas==2.0.2
scikit-learn==1.3.2
matplotlib==3.7.2
tqdm==4.66.3 